backend/requirements.txt
coverage==4.5.3
Django==2.1.7
django-cors-headers==2.5.2
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
PyJWT==1.7.1
pytz==2018.9
